+++ resources/views/airportLink/overview.blade.php
@@ -4,7 +4,7 @@
     <h2>Airport Links Overview</h2>
     @include('layouts.alert')
     <p>
-        <a href=" {{ route('airportLink.create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Add new Airport
+        <a href=" {{ route('airportLinks.create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Add new Airport
             Link</a>
     </p>
     <table class="table table-hover">
@@ -17,17 +17,17 @@
         </thead>
         @forelse($airportLinks as $airportLink)
             <tr>
-                <td><a href="{{ route('airport.show', $airportLink->airport) }}">{{ $airportLink->airport->name }} [{{ $airportLink->airport->icao }} | {{ $airportLink->airport->iata }}]</a></td>
+                <td><a href="{{ route('airports.show', $airportLink->airport) }}">{{ $airportLink->airport->name }} [{{ $airportLink->airport->icao }} | {{ $airportLink->airport->iata }}]</a></td>
                 <td><a href="{{ $airportLink->url }}">{{ $airportLink->name ?? $airportLink->type->name }}</a></td>
                 <td>
-                    <a href="{{ route('airportLink.edit', $airportLink) }}">
+                    <a href="{{ route('airportLinks.edit', $airportLink) }}">
                         <button class="btn btn-primary">
                             <i class="fa fa-edit"></i> Edit Airport Link
                         </button>
                     </a>
                 </td>
                 <td>
-                    <form action="{{ route('airportLink.destroy', $airportLink) }}" method="post">
+                    <form action="{{ route('airportLinks.destroy', $airportLink) }}" method="post">
                         @method('DELETE')
                         <button class="btn btn-danger"><i class="fa fa-trash"></i> Remove Airport Link</button>
                         @csrf