+++ resources/views/event/admin/import.blade.php
@@ -17,7 +17,7 @@
                 <div class="card-header">{{ $event->name }} | Import</div>
 
                 <div class="card-body">
-                    <form method="POST" action="{{ route('booking.admin.import',$event) }}"
+                    <form method="POST" action="{{ route('bookings.admin.import',$event) }}"
                           enctype="multipart/form-data">
                         @csrf
                         @method('PUT')