+++ resources/views/event/overview.blade.php
@@ -2,7 +2,7 @@
 
 @section('content')
     <h2>Events Overview</h2>
-    <p><a href="{{ route('event.create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Add new Event</a></p>
+    <p><a href="{{ route('events.create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Add new Event</a></p>
     @include('layouts.alert')
     <table class="table table-hover">
         <thead>
@@ -23,30 +23,30 @@
                 <tr>
                     @endif
                     <td>{{ $event->id }}</td>
-                    <td><a href="{{ route('booking.index', $event) }}" title="Open Slot Table">{{ $event->name }}</a>
+                    <td><a href="{{ route('bookings.event.index', $event) }}" title="Open Slot Table">{{ $event->name }}</a>
                     </td>
                     <td>{{ $event->type->name }}</td>
                     <td>{{ $event->startEvent->format('d-m-Y') }}</td>
                     <td>{{ $event->startEvent->format('Hi') }}z</td>
                     <td>{{ $event->endEvent->format('Hi') }}z</td>
                     <td>
-                        <a href="{{ route('event.edit',$event->id) }}" role="button" class="btn btn-primary disabled"><i
+                        <a href="{{ route('events.edit',$event->id) }}" role="button" class="btn btn-primary disabled"><i
                                     class="fa fa-edit"></i> Edit Event</a>&nbsp;
                         @if($event->endEvent > now())
-                            <a href="{{ route('booking.admin.importForm',$event) }}" class="btn btn-success"><i
+                            <a href="{{ route('bookings.admin.importForm',$event) }}" class="btn btn-success"><i
                                         class="fa fa-edit"></i> Import data</a>&nbsp;
-                            <a href="{{ route('booking.create',$event) }}" class="btn btn-primary"><i
+                            <a href="{{ route('bookings.create',$event) }}" class="btn btn-primary"><i
                                         class="fa fa-plus"></i> Add Timeslots</a>&nbsp;
-                            <a href="{{ route('booking.admin.autoAssignForm',$event) }}" class="btn btn-primary">Auto
+                            <a href="{{ route('bookings.admin.autoAssignForm',$event) }}" class="btn btn-primary">Auto
                                 Assign FL / Route</a>&nbsp;
                             @if($event->is_oceanic_event)
-                                <a href="{{ route('booking.admin.autoAssignForm',$event) }}" class="btn btn-primary">
+                                <a href="{{ route('bookings.admin.autoAssignForm',$event) }}" class="btn btn-primary">
                                     Auto Assign FL / Route</a>&nbsp;
                             @endif
                         @endif
-                        <a href="{{ route('event.email.form',$event) }}" class="btn btn-primary"><i
+                        <a href="{{ route('events.email.form',$event) }}" class="btn btn-primary"><i
                                     class="fa fa-envelope"></i> Send E-mails (all persons)</a>&nbsp;
-                        <a href="{{ route('booking.export',$event) }}" class="btn btn-success"><i
+                        <a href="{{ route('bookings.export',$event) }}" class="btn btn-success"><i
                                     class="fa fa-edit"></i> Export data</a>&nbsp;
                     </td>
                 </tr>