+++ resources/views/event/sendEmail.blade.php
@@ -18,7 +18,7 @@
                 <div class="card-header">{{ $event->name }} | Send Bulk E-mail</div>
 
                 <div class="card-body">
-                    <form method="POST" action="{{ route('event.email',$event) }}">
+                    <form method="POST" action="{{ route('events.email',$event) }}">
                         @csrf
                         @method('PATCH')
 
@@ -27,7 +27,7 @@
                             <label class="col-md-4 col-form-label text-md-right"></label>
 
                             <div class="col-md-6">
-                                <a href="{{ route('event.email.final',$event) }}" class="btn btn-primary"><i
+                                <a href="{{ route('events.email.final',$event) }}" class="btn btn-primary"><i
                                             class="fa fa-envelope"></i> Send Final Information E-mail</a>
                             </div>
                         </div>